预案bug修改、事件视频 获取上游相机优化

+++ tunnel-business/src/main/java/com/tunnel/business/service/event/impl/SdReserveProcessServiceImpl.java
@@ -104,19 +104,23 @@ public class SdReserveProcessServiceImpl implements ISdReserveProcessService {
     @Override
     @Transactional(rollbackFor = Exception.class)
     public int batchSdReserveProcessed(SdReserveProcessModel sdReserveProcesses) {
+        if(sdReserveProcesses.getSdReserveProcesses().isEmpty()){
+            throw new RuntimeException("无效数据策略添加失败");
+        }
         List<SdReserveProcess> list = new ArrayList<>();
         String planId = sdReserveProcesses.getReserveId().toString();
         //删除预案流程节点
         sdReserveProcessMapper.deleteSdReserveProcessByPlanId(sdReserveProcesses.getReserveId());
         SpringUtils.getBean(SdStrategyRlMapper.class).deleteSdStrategyRlByPlanId(Long.valueOf(planId));
         for (Map process : sdReserveProcesses.getSdReserveProcesses()) {
+            List<String> value = (List<String>) process.get("equipments");
+            if(value.isEmpty() || process.get("state").equals("")){
+                throw new RuntimeException("无效数据策略添加失败");
+            }
             SdReserveProcess reserveProcess = new SdReserveProcess();
             SdStrategyRl rl = new SdStrategyRl();
             String equipments = "";
-            if(StrUtil.isNotBlank(process.get("equipments").toString())){
-                List<String> value = (List<String>) process.get("equipments");
-                equipments = StringUtils.join(value,",");
-            }
+            equipments = StringUtils.join(value,",");
             String equipmentTypeId = process.get("eqTypeId") + "";
             String eqState = (String) process.get("state");
             rl.setEqTypeId(equipmentTypeId);
@@ -125,31 +129,14 @@ public class SdReserveProcessServiceImpl implements ISdReserveProcessService {
             rl.setPlanId(planId);
             rl.setRetrievalRule(process.get("retrievalRule").toString());
             SpringUtils.getBean(SdStrategyRlMapper.class).insertSdStrategyRl(rl);
-            //rl.getId();
             reserveProcess.setProcessName(process.get("processName").toString());
             reserveProcess.setProcessSort(Integer.parseInt(process.get("processSort").toString()));
             reserveProcess.setDeviceTypeId(Long.valueOf(equipmentTypeId));
             reserveProcess.setStrategyId(rl.getId());
             reserveProcess.setReserveId(Long.valueOf(planId));
-            //SpringUtils.getBean(SdStrategyRlMapper.class).insertSdStrategyRl(rl);
-//            Long[] strategyIds = process.getHandleStrategyList();
-//            List<SdStrategyRl> rlList = SpringUtils.getBean(SdStrategyRlMapper.class).selectSdStrategyRlByStrategyId(strategyIds[1]);
-//            if(rlList.isEmpty()){
-//                continue;
-//            }
-//            reserveProcess.setReserveId(sdReserveProcesses.getReserveId());
-//            reserveProcess.setDeviceTypeId(Long.parseLong(rl.getEqTypeId()));
-//            reserveProcess.setProcessName(process.getProcessName());
-//            reserveProcess.setProcessSort(process.getProcessSort());
-//            reserveProcess.setStrategyId(rl.getId());
-//            reserveProcess.setCreateTime(DateUtils.getNowDate());
-//            reserveProcess.setCreateBy(SecurityUtils.getUsername());
             list.add(reserveProcess);
         }
         int result = -1;
-        if(list.isEmpty()){
-            throw new RuntimeException("无效数据策略添加失败");
-        }
         result = sdReserveProcessMapper.batchSdReserveProcess(list);
         return result;
     }
